+++ .github/workflows/opcua-plc.yml
@@ -25,11 +25,11 @@ env:
   GO_VERSION: '1.21.*'
 
 concurrency:
-  group: wago-s7-test
+  group: opcua-plc-test
   cancel-in-progress: true
 
 jobs:
-  go-test-wago-s7:
+  go-test-opcua-plc:
     runs-on: hercules
     permissions:
       packages: write
