+++ .github/workflows/s7-plc.yml
@@ -25,11 +25,11 @@ env:
   GO_VERSION: '1.21.*'
 
 concurrency:
-  group: s7-test
+  group: s7-plc-test
   cancel-in-progress: true
 
 jobs:
-  go-test-s7:
+  go-test-s7-plc:
     runs-on: hercules
     timeout-minutes: 30
     steps:
