+++ opcua_plugin/opcua_simulator_test.go
@@ -312,51 +312,3 @@ var _ = Describe("Test Against Microsoft OPC UA simulator", func() {
 	// TODO: Move more tests over from TestAgainstSimulator in opcua_test.go
 
 })
-
-var _ = FDescribe("Test Against Siemens S7", func() {
-
-	var endpoint string
-
-	BeforeEach(func() {
-		endpoint = os.Getenv("TEST_S7_ENDPOINT_URI")
-
-		// Check if environment variables are set
-		if endpoint == "" {
-			Skip("Skipping test: environment variables are not set")
-			return
-		}
-	})
-
-	Describe("Connect", func() {
-		It("should connect", func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			nodeIDStrings := []string{"ns=4;i=1"}
-
-			parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-			input := &OPCUAInput{
-				Endpoint:         endpoint,
-				Username:         "",
-				Password:         "",
-				NodeIDs:          parsedNodeIDs,
-				SubscribeEnabled: false,
-			}
-
-			// Attempt to connect
-			err := input.Connect(ctx)
-			Expect(err).NotTo(HaveOccurred())
-
-			messageBatch, _, err := input.ReadBatch(ctx)
-			Expect(err).NotTo(HaveOccurred())
-
-			Expect(messageBatch).To(HaveLen(3))
-
-			// Close connection
-			if input.Client != nil {
-				input.Client.Close(ctx)
-			}
-		})
-	})
-})
